+++ app/models/jupiter_core/search.rb
@@ -37,8 +37,12 @@ class JupiterCore::Search
 
     base_query << q if q.present?
     facets.each do |key, values|
-      values.each do |value|
-        fq << %Q(#{key}: "#{value}")
+      if Flipper.enabled?(:or_facets)
+        fq << %Q(#{key}:\(#{values.collect { |value| "\"#{value}\"" }.join(' OR ')}\))
+      else
+        values.each do |value|
+          fq << %Q(#{key}:"#{value}")
+        end
       end
     end
     ranges.each do |key, value|